server/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0
supabase>=2.0.0
pypdf>=3.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# Testing dependencies
pytest>=8.0.0
pytest-asyncio>=0.20.0
pytest-cov>=4.0.0
httpx>=0.25.0

# Development and linting dependencies
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.0.0